Begin building the chat interface
A user can enter their name and a message into a form and send it off to
the server for "processing". What is meant by "processing" will be
refined over each iteration. At this point "processing" just means the
server is able to retrieve the data sent by the user.